Now storage and cache are aware of each other, and the organizer is passed a copy.
Then, I implemented cold cache hits (which would query the storage for
permanent information). This way, the cache doesn't need to be fully
loaded on start, but will be warmed up by certain hits.

Additionally, fixed the path location for where to fetch the package. Next
step: write fetch package (for install purposes).

The registry is mostly complete now, save for some dependency tracking
things, and the future S3 enhancements. Once is stored and retrieves
packages with authentication, that's pretty much it.